Update testcase for invalid data
