+++ tests/api/main.rs
@@ -1,3 +1,3 @@
 mod health_check;
 mod helpers;
-mod users;
+mod post_user;
